+++ project/app/src/main/java/org/owntracks/android/ui/AppShortcuts.kt
@@ -3,6 +3,5 @@ package org.owntracks.android.ui
 import android.content.Context
 
 interface AppShortcuts {
-    fun enableLogViewerShortcut(applicationContext: Context)
-    fun disableLogViewerShortcut(applicationContext: Context)
+    fun enableShortcuts(applicationContext: Context)
 }