+++ project/app/src/test/java/org/owntracks/android/preferences/PreferenceTest.kt
@@ -10,7 +10,6 @@ import org.junit.Test
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
 import org.owntracks.android.model.messages.MessageConfiguration
-import org.owntracks.android.ui.NoopAppShortcuts
 
 class PreferenceTest {
     private lateinit var mockResources: Resources
@@ -27,14 +26,14 @@ class PreferenceTest {
 
     @Test
     fun `given a single key value, when importing to preferences, then that value can be retrieved from the preferences`() {
-        val preferences = Preferences(mockContext, preferencesStore, NoopAppShortcuts())
+        val preferences = Preferences(preferencesStore)
         preferences.importKeyValue("ignoreStaleLocations", 195.4f)
         assertEquals(195.4f, preferences.ignoreStaleLocations, 0.001f)
     }
 
     @Test
     fun `given a configuration message, when importing to preferences, all the keys in the config should be added`() {
-        val preferences = Preferences(mockContext, preferencesStore, NoopAppShortcuts())
+        val preferences = Preferences(preferencesStore)
         val messageConfiguration = MessageConfiguration()
         messageConfiguration["autostartOnBoot"] = true
         messageConfiguration["host"] = "testhost"
@@ -47,7 +46,7 @@ class PreferenceTest {
 
     @Test
     fun `given a configuration message with an entry value of null, when importing to preferences, the config value should be cleared`() {
-        val preferences = Preferences(mockContext, preferencesStore, NoopAppShortcuts())
+        val preferences = Preferences(preferencesStore)
         val messageConfiguration = MessageConfiguration()
         messageConfiguration["host"] = null
         preferences.host = "testHost"
@@ -57,7 +56,7 @@ class PreferenceTest {
 
     @Test
     fun `given a configuration message with an invalid key, when importing to preferences, it should be ignored`() {
-        val preferences = Preferences(mockContext, preferencesStore, NoopAppShortcuts())
+        val preferences = Preferences(preferencesStore)
         val messageConfiguration = MessageConfiguration()
         messageConfiguration["Invalid"] = "invalid"
         preferences.importConfiguration(messageConfiguration)
@@ -66,7 +65,7 @@ class PreferenceTest {
 
     @Test
     fun `given a configuration message with a value of the wrong type, when importing to preferences, it should be ignored`() {
-        val preferences = Preferences(mockContext, preferencesStore, NoopAppShortcuts())
+        val preferences = Preferences(preferencesStore)
         val messageConfiguration = MessageConfiguration()
         messageConfiguration["host"] = 4
         preferences.importConfiguration(messageConfiguration)
@@ -124,7 +123,7 @@ class PreferenceTest {
 
     @Test
     fun `given an MQTT configuration message, when imported and then exported, the config is merged and all the preference keys are present`() {
-        val preferences = Preferences(mockContext, preferencesStore, NoopAppShortcuts())
+        val preferences = Preferences(preferencesStore)
         val messageConfiguration = MessageConfiguration()
         messageConfiguration["autostartOnBoot"] = true
         messageConfiguration["host"] = "testhost"
@@ -155,7 +154,7 @@ class PreferenceTest {
 
     @Test
     fun `given an HTTP configuration message, when imported and then exported, the config is merged and all the preference keys are present`() {
-        val preferences = Preferences(mockContext, preferencesStore, NoopAppShortcuts())
+        val preferences = Preferences(preferencesStore)
         val messageConfiguration = MessageConfiguration()
         messageConfiguration["autostartOnBoot"] = true
         messageConfiguration["host"] = "testhost"
@@ -186,7 +185,7 @@ class PreferenceTest {
 
     @Test
     fun `given a Preferences object with no username set, when asking for the topic, the correct username placeholder is populated`() {
-        val preferences = Preferences(mockContext, preferencesStore, NoopAppShortcuts())
+        val preferences = Preferences(preferencesStore)
         preferences.autostartOnBoot = true
         preferences.username = ""
         preferences.deviceId = "myDevice"
@@ -196,7 +195,7 @@ class PreferenceTest {
 
     @Test
     fun `given an empty Preferences object, when asking for a value, then the default value is returned`() {
-        val preferences = Preferences(mockContext, preferencesStore, NoopAppShortcuts())
+        val preferences = Preferences(preferencesStore)
         assertEquals(false, preferences.debugLog)
     }
 }